requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pyyaml==6.0.2
cachetools==5.5.2
python-dotenv==1.0.1
cryptography==44.0.2
